refactor: add useWaypoint hook
